requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
PyMySQL==1.1.1
gunicorn==22.0.0
python-dotenv==1.0.1
validators==0.34.0